prediction_model/macroeconomic/requirements_ml.txt
tensorflow>=2.16.0
keras>=3.0.0
nltk>=3.8.0
scikit-learn>=1.4.0
beautifulsoup4>=4.12.0
seaborn>=0.13.0
requests>=2.32.0
tavily-python>=0.7.14
python-dotenv>=1.2.1
feedparser>=6.0.10
selenium>=4.18.0
webdriver-manager>=4.0.1
google-generativeai>=0.4.0
sanpy
scipy
